backport r645 to 14.08
3.10 kernel: gpio: pca953x: backport gpio input fix

This fix is for anyone on 3.10 using the pca953x driver. A regression
was introduced when this driver was converted to using 8-bit
reads/writes the bitmask in pca953x_gpio_get_value wasn't adjusted with
the modulus BANK_SZ and consequently looks at the wrong bits in the
input register.

This backports the following commit from the mainline linux kernel:
40a625daa88653d7942dc85483f6f289cd687cb7

Signed-off-by: Pushpal Sidhu <psidhu@gateworks.com>


git-svn-id: http://svn.gateworks.com/openwrt@646 918123d7-9a9f-4a10-9424-6605889f5124