laguna: add GSC drivers for watchdog and input
The GSC has recently gained hardware watchdog support which is unique in that
the expiration of this watchdog will truely power cycle the board vs doing
a chip level reset.

This patch adds a Linux watchdog driver which will
take priority over the ARM mpcore watchdog inside the cavium if the
GSC register configuration enables the hardware watchdog and firmware v44
or greater is on the GSC (otherwise the ARM mpcore watchdog will be used).

This patch additionally adds a driver that handles the various interrupt
sources from the GSC and connects the user pushbutton interrupt to a hotplug
event through OpenWrt's gpio-button-hotplug driver.

See http://trac.gateworks.com/wiki/gsc#GSCDrivers for more details

Signed-off-by: Tim Harvey <tharvey@gateworks.com>


git-svn-id: http://svn.gateworks.com/openwrt@694 918123d7-9a9f-4a10-9424-6605889f5124