backport r731 to 14.08
ventana: use OTP values to set imx-thermal driver thresholds

The IMX6Q/IMX6DL SoC's have a 2-bit temperature grade stored in OTP which
is valid for all IMX6 SoC's (despite the fact that the IMXSDLRM and
IMXSXRM do not document this - this has been proven via tests as well as
verified by Freescale FAE).

Instead of assuming a fixed 85C for passive cooling threshold and 105C for
critical use the thermal grade for these configurations.

We will set the critical to maxT - 5C and passive to maxT - 10C.

Signed-off-by: Tim Harvey <​tharvey@gateworks.com>


git-svn-id: http://svn.gateworks.com/openwrt@732 918123d7-9a9f-4a10-9424-6605889f5124