Ventana/Laguna: hwmon: gsc: improve sensor reading robustness
The GSC can occasionally return a failure if the device is busy in an ADC
cycle. Therefore we retry up to 3 times (which provides enough delay for
any pending ADC to complete). Additionally, if some error does occur, return
the error value instead of allowing it to corrupt one byte of the multi-byte
reading.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>


git-svn-id: http://svn.gateworks.com/openwrt@747 918123d7-9a9f-4a10-9424-6605889f5124