backport r769 to 14.08:
Laguna: add gsc-input driver

Update the gsc-input driver to support non device-tree kernels and add the
event_test package to the Laguna filesystem.

This provides a mechanism to easily get interrupt driven events for gsc
interrupts such as pushbutton events, tamper events, etc.

Note that the gpio-button-hotplug driver is also installed which presents
the user pushbutton press events and release events as a hotplug event which
does not require the gsc-input driver.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>

git-svn-id: http://svn.gateworks.com/openwrt@770 918123d7-9a9f-4a10-9424-6605889f5124